requirements.txt
streamlit==1.45.0
psycopg2-binary==2.9.10
plotly==5.24.1
pandas==2.2.3
httpx==0.28.1
pydantic-settings==2.7.0
python-dotenv==1.0.1
PyYAML==6.0.2
tabulate==0.9.0
json-repair==0.30.3
